@@ -9,15 +9,15 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.47",
-    "@types/react": "^16.14.24",
-    "@types/react-dom": "^16.9.14",
-    "@types/react-redux": "^7.1.23",
+    "@types/node": "^12.20.50",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-redux": "^7.1.24",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-redux": "^7.2.7",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^7.2.8",
     "react-scripts": "5.0.0",
     "typescript": "~4.1.5"
   },
